WIP: GCSE details controller
