+++ app/controllers/candidate_interface/gcse/base_controller.rb
@@ -1,0 +1,22 @@
+module CandidateInterface
+  class Gcse::BaseController < CandidateInterfaceController
+    before_action :redirect_to_dashboard_if_submitted
+    before_action :set_subject
+    before_action :render_application_feedback_component
+
+  private
+
+    def set_subject
+      @subject = subject_param
+    end
+
+    def subject_param
+      params.require(:subject)
+    end
+
+    def update_gcse_completed(value)
+      attribute_to_update = "#{@subject}_gcse_completed"
+      current_application.update!("#{attribute_to_update}": value)
+    end
+  end
+end
