+++ app/controllers/candidate_interface/gcse/year_controller.rb
@@ -1,6 +1,5 @@
 module CandidateInterface
-  class Gcse::YearController < Gcse::DetailsController
-    before_action :redirect_to_dashboard_if_submitted
+  class Gcse::YearController < Gcse::BaseController
 
     def update
       @qualification_type = details_form.qualification.qualification_type
